// frontend/src/app/models/estate.ts
export class Estate {
    id: number;
    type: string;
    name: string;
    city: string;
    municipality: string;
    location: string;
    street: string;
    area: number;
    rooms: number;
    constructionYear: number;
    state: string;
    heating: string;
    floor: number;
    totalFloors: number;
    parking: string;
    monthly: number;
    price: number;
    about: string;
    characteristics: string[];
    busses: number[];
    agencyName: string;
    agencyPIB: string;
    agencyCity: string;
    agencyAddress: string;
    agencyPhone: string;
    adFirstName: string;
    adSurname: string;
    adPhone: string;
    adLicense: string;
    sold: number;
    lastModified: string;
    images: string[];
    lat:number;
    lng:number;
}